# packages/contracts/openapi/control-v1.yaml
...
openapi: 3.1.0
info:
  title: OLO ToolGate Control Plane API
  version: 0.10.0-dev
  description: Tenant-scoped group administration with reviewed configuration changes
    and online current authorization. Signed snapshots describe configuration; execution
    requires a separately signed single-use permit.
security:
- adminAccessToken: []
paths:
  /api/control/v1/admin-session:
    get:
      operationId: getAdminSession
      responses:
        '200':
          description: Verified administrative role
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/control.schema.json#/$defs/AdminSession
        '403':
          description: Administrator portal permission required
...
  /api/control/v1/bundles/current:
    get:
      operationId: getCurrentPolicyBundle
      description: Tenant-scoped immutable signed current bundle. Requires admin,
        reader or dedicated toolgate-bundle-reader. No-store; ETag is sequence. Refresh
        authentication externally.
      responses:
        '200':
          description: Signed bundle; clients must verify signature, hash, tenant,
            trust claims and freshness before adoption
          headers:
            ETag:
              schema:
                type: string
              description: Quoted immutable sequence
            X-Request-ID:
              schema:
                type: string
              description: Server correlation ID
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/bundle.schema.json#/$defs/SignedPolicyBundle
        '401':
          description: Valid short-lived access token required
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/error.schema.json#/$defs/ErrorEnvelope
        '403':
          description: Role forbidden
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/error.schema.json#/$defs/ErrorEnvelope
        '404':
          description: No publication or historical version in this tenant
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/error.schema.json#/$defs/ErrorEnvelope
        '503':
          description: Signing disabled, signing failure or database unavailable
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/error.schema.json#/$defs/ErrorEnvelope
  /api/control/v1/bundles/versions/{sequence}:
    get:
      operationId: getPolicyBundleVersion
      description: Exact immutable historical bytes. Requires admin or reader. Historical
        expiry is not extended.
      parameters:
      - name: sequence
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          maximum: 9007199254740991
      responses:
        '200':
          description: Signed bundle; clients must verify signature, hash, tenant,
            trust claims and freshness before adoption
          headers:
            ETag:
              schema:
                type: string
              description: Quoted immutable sequence
            X-Request-ID:
              schema:
                type: string
              description: Server correlation ID
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/bundle.schema.json#/$defs/SignedPolicyBundle
        '401':
          description: Valid short-lived access token required
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/error.schema.json#/$defs/ErrorEnvelope
        '403':
          description: Role forbidden
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/error.schema.json#/$defs/ErrorEnvelope
        '404':
          description: No publication or historical version in this tenant
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/error.schema.json#/$defs/ErrorEnvelope
        '503':
          description: Signing disabled, signing failure or database unavailable
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/error.schema.json#/$defs/ErrorEnvelope
  /api/control/v1/bundles/publish:
    post:
      operationId: publishPolicyBundle
      description: "Admin only. Compile current directoryRevision and atomically publish\
        \ expectedSequence+1 with audit and durable replay. rollbackOf must be absent.\
        \ Lifetime is 1 second\xC3\u0192\xC6\u2019\xC3\u201A\xC2\xA2\xC3\u0192\xC2\
        \xA2\xC3\xA2\xE2\u201A\xAC\xC5\xA1\xC3\u201A\xC2\xAC\xC3\u0192\xC2\xA2\xC3\
        \xA2\xE2\u20AC\u0161\xC2\xAC\xC3\u2026\xE2\u20AC\u015324 hours; grace is 0\xC3\
        \u0192\xC6\u2019\xC3\u201A\xC2\xA2\xC3\u0192\xC2\xA2\xC3\xA2\xE2\u201A\xAC\
        \xC5\xA1\xC3\u201A\xC2\xAC\xC3\u0192\xC2\xA2\xC3\xA2\xE2\u20AC\u0161\xC2\xAC\
        \xC3\u2026\xE2\u20AC\u01535 minutes. Publication does not execute tools."
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          $ref: ../schemas/v1/identifiers.schema.json#/$defs/Identifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../schemas/v1/bundle.schema.json#/$defs/BundlePublishRequest
      responses:
        '201':
          description: Signed bundle; clients must verify signature, hash, tenant,
            trust claims and freshness before adoption
          headers:
            ETag:
              schema:
                type: string
              description: Quoted immutable sequence
            X-Request-ID:
              schema:
                type: string
              description: Server correlation ID
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/bundle.schema.json#/$defs/SignedPolicyBundle
        '400':
          description: Malformed or unsupported input; required fields/key missing
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/error.schema.json#/$defs/ErrorEnvelope
        '401':
          description: Valid short-lived access token required
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/error.schema.json#/$defs/ErrorEnvelope
        '403':
          description: Role forbidden
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/error.schema.json#/$defs/ErrorEnvelope
        '404':
          description: No publication or historical version in this tenant
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/error.schema.json#/$defs/ErrorEnvelope
        '409':
          description: Stale directory/bundle sequence or changed idempotency request
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/error.schema.json#/$defs/ErrorEnvelope
        '503':
          description: Signing disabled, signing failure or database unavailable
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/error.schema.json#/$defs/ErrorEnvelope
  /api/control/v1/bundles/rollback:
    post:
      operationId: rollbackPolicyBundle
      description: Admin only. rollbackOf is required and selects an existing tenant
        version. Copies its exact policy bytes into expectedSequence+1 with fresh
        signed expiry, new signature, audit and replay. Does not rewind sequence or
        mutate history.
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          $ref: ../schemas/v1/identifiers.schema.json#/$defs/Identifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../schemas/v1/bundle.schema.json#/$defs/BundlePublishRequest
      responses:
        '201':
          description: Signed bundle; clients must verify signature, hash, tenant,
            trust claims and freshness before adoption
          headers:
            ETag:
              schema:
                type: string
              description: Quoted immutable sequence
            X-Request-ID:
              schema:
                type: string
              description: Server correlation ID
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/bundle.schema.json#/$defs/SignedPolicyBundle
        '400':
          description: Malformed or unsupported input; required fields/key missing
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/error.schema.json#/$defs/ErrorEnvelope
        '401':
          description: Valid short-lived access token required
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/error.schema.json#/$defs/ErrorEnvelope
        '403':
          description: Role forbidden
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/error.schema.json#/$defs/ErrorEnvelope
        '404':
          description: No publication or historical version in this tenant
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/error.schema.json#/$defs/ErrorEnvelope
        '409':
          description: Stale directory/bundle sequence or changed idempotency request
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/error.schema.json#/$defs/ErrorEnvelope
        '503':
          description: Signing disabled, signing failure or database unavailable
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/error.schema.json#/$defs/ErrorEnvelope
  /api/control/v1/openapi:
    get:
      operationId: getControlOpenApi
      description: Requires toolgate-admin or toolgate-reader and valid tenant identity
        claims. Returns this API with all schemas embedded.
      responses:
        '200':
          description: Self-contained OpenAPI 3.1 document
          content:
            application/json:
              schema:
                type: object
        '401':
          description: Verified access token and valid identity claims required
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/error.schema.json#/$defs/ErrorEnvelope
        '403':
          description: Role forbidden
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/error.schema.json#/$defs/ErrorEnvelope
...
  /api/control/v1/config/export:
    get:
      operationId: exportConfig
      description: Requires toolgate-admin or toolgate-reader.
      responses:
        '200':
          description: Success
          headers:
            ETag:
              description: Quoted revision
              schema:
                type: string
            X-Request-ID:
              description: Server correlation ID
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/control.schema.json#/$defs/ControlSnapshot
            application/yaml:
              schema:
                $ref: ../schemas/v1/control.schema.json#/$defs/ControlSnapshot
        '400': *id002
        '401': *id003
        '403': *id004
        '404': *id005
        '409': *id006
        '413': *id007
        '415': *id008
        '500': *id009
        '503': *id010
  /api/control/v1/config/import:
    post:
      operationId: importConfig
      description: Requires toolgate-admin.
      responses:
        '200':
          description: Success
          headers:
            ETag:
              description: Quoted revision
              schema:
                type: string
            X-Request-ID:
              description: Server correlation ID
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/control.schema.json#/$defs/ControlImportResult
        '400': *id002
        '401': *id003
        '403': *id004
        '404': *id005
        '409': *id006
        '413': *id007
        '415': *id008
        '500': *id009
        '503': *id010
        '202':
          description: Draft created; review impact, submit for independent review,
            then apply
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterpriseConfigurationChange
      parameters:
      - *id011
      - *id012
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../schemas/v1/control.schema.json#/$defs/ControlImportRequest
          application/yaml:
            schema:
              $ref: ../schemas/v1/control.schema.json#/$defs/ControlImportRequest
  /api/control/v1/audit:
    get:
      operationId: listAudit
      description: Requires toolgate-admin.
      responses:
        '200':
          description: Success
          headers:
            ETag:
              description: Quoted revision
              schema:
                type: string
            X-Request-ID:
              description: Server correlation ID
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/control.schema.json#/$defs/ControlAuditPage
        '400': *id002
        '401': *id003
        '403': *id004
        '404': *id005
        '409': *id006
        '413': *id007
        '415': *id008
        '500': *id009
        '503': *id010
      parameters:
      - name: cursor
        in: query
        schema:
          type: integer
          minimum: 0
          default: 0
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 50
...
  /api/control/v1/configuration-changes:
    get:
      operationId: listConfigurationChanges
      parameters:
      - name: after
        in: query
        schema:
          type: string
          maxLength: 128
      responses:
        '200':
          description: Current reviewed configuration state
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterpriseConfigurationPage
        '403':
          description: Current scoped authority required
        '409':
          description: Revision, digest or epoch changed
  /api/control/v1/configuration-changes/{id}:
    get:
      operationId: getConfigurationChange
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          maxLength: 128
      responses:
        '200':
          description: Current reviewed configuration state
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterpriseConfigurationChange
        '403':
          description: Current scoped authority required
        '409':
          description: Revision, digest or epoch changed
  /api/control/v1/configuration-changes/{id}/transition:
    post:
      operationId: transitionConfigurationChange
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          maxLength: 128
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterpriseConfigurationTransition
      responses:
        '200':
          description: Current reviewed configuration state
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterpriseConfigurationChange
        '403':
          description: Current scoped authority required
        '409':
          description: Revision, digest or epoch changed
...
components:
  securitySchemes:
    adminAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 token with verified issuer/audience/expiry and signed tenant_id/groups.
        Not a Gateway runtime credential.
    deviceCertificate:
      type: mutualTLS
      description: Direct verified device TLS identity; forwarded identity headers
        never authenticate.